Added NumberContainer in CPAlien
git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/explicitfloat@10219 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c

+++ src/org/sosy_lab/cpachecker/cpa/cpalien/SMGExpressionEvaluator.java
@@ -267,7 +267,7 @@ public class SMGExpressionEvaluator {
     if (value == null) {
       return SMGUnknownValue.getInstance();
     } else {
-      return SMGKnownExpValue.valueOf(value.longValue());
+      return SMGKnownExpValue.valueOf(value);
     }
   }
