+++ src/org/sosy_lab/cpachecker/cpa/explicit/SMGExplicitCommunicator.java
@@ -354,7 +354,7 @@ public class SMGExplicitCommunicator {
       if (value == null) {
         return SMGUnknownValue.getInstance();
       } else {
-        return SMGKnownExpValue.valueOf(value.longValue());
+        return SMGKnownExpValue.valueOf(value);
       }
     }
   }
